municipalty-project/backend/requirement.txt
fastapi
python-decouple
pydantic
pymongo
python-jose[Cryptography]
python-multipart
python-dotenv
uvicorn[standard]
passlib[bcrpyt]